build: move gulp tasks and deps to new gulp-tasks package
