+++ infra/gulp-tasks/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@markdown-editor/gulp-tasks",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "files": [
+    "index",
+    "build"
+  ],
+  "exports": {
+    ".": "./index.js",
+    "./build": "./build.js"
+  },
+  "dependencies": {
+    "@gravity-ui/gulp-utils": "1.0.3",
+    "@types/gulp": "4.0.9",
+    "@types/gulp-sass": "5.0.0",
+    "gulp": "5.0.0",
+    "gulp-concat": "2.6.1",
+    "gulp-replace": "1.1.3",
+    "gulp-sass": "6.0.0",
+    "gulp-sourcemaps": "3.0.0",
+    "rimraf": "3.0.2",
+    "sass": "^1.84.0"
+  },
+  "devDependencies": {
+    "typescript": "catalog:ts"
+  },
+  "peerDependencies": {
+    "typescript": ">=5"
+  }
+}
