+++ packages/editor/gulpfile.mjs
@@ -1,0 +1,20 @@
+import {dirname, resolve} from 'node:path';
+import {fileURLToPath} from 'node:url';
+
+import {series, task} from '@markdown-editor/gulp-tasks';
+import {registerBuildTasks} from '@markdown-editor/gulp-tasks/build';
+
+import pkg from './package.json' with {type: 'json'};
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
+
+const BUILD_DIR = resolve('build');
+const NODE_MODULES_DIR = resolve(__dirname, 'node_modules');
+
+registerBuildTasks({
+    version: pkg.version,
+    buildDir: BUILD_DIR,
+    nodeModulesDir: NODE_MODULES_DIR,
+});
+
+task('default', series('clean', 'build'));
