+++ packages/editor/package.json
@@ -223,10 +223,7 @@
     "@diplodoc/themes": "^1.0.0",
     "@diplodoc/transform": "catalog:peer-diplodoc",
     "@gravity-ui/components": "catalog:peer-gravity",
-    "@gravity-ui/gulp-utils": "1.0.3",
     "@gravity-ui/uikit": "catalog:peer-gravity",
-    "@types/gulp": "4.0.9",
-    "@types/gulp-sass": "5.0.0",
     "@types/jest": "^29.5.14",
     "@types/jsdom": "21.1.7",
     "@types/katex": "0.16.7",
@@ -236,17 +233,13 @@
     "@types/react-dom": "catalog:react",
     "@types/rimraf": "3.0.2",
     "@types/sanitize-html": "2.11.0",
+    "@markdown-editor/gulp-tasks": "workspace:*",
     "@markdown-editor/tsconfig": "workspace:*",
     "bem-cn-lite": "4.1.0",
     "dpdm": "3.14.0",
     "esbuild": "0.21.5",
     "esbuild-sass-plugin": "2.15.0",
-    "gulp": "5.0.0",
     "gulp-cli": "3.0.0",
-    "gulp-concat": "2.6.1",
-    "gulp-replace": "1.1.3",
-    "gulp-sass": "6.0.0",
-    "gulp-sourcemaps": "3.0.0",
     "highlight.js": "catalog:peers",
     "identity-obj-proxy": "^3.0.0",
     "ist": "1.1.7",
@@ -264,9 +257,6 @@
     "prettier": "catalog:linters",
     "react": "catalog:react",
     "react-dom": "catalog:react",
-    "rimraf": "3.0.2",
-    "sass": "^1.84.0",
-    "sass-loader": "^13.3.2",
     "ts-jest": "^29.2.5",
     "typescript": "catalog:ts"
   },
